All tests passing after refactor
