+++ tests/test_run.py
@@ -1,6 +1,6 @@
 """ Test the script that runs (calls Proctor) benchmarks on LLMs """
 from unittest import mock
-import run
+from tether.core.run import main
 
 def test_run_main(monkeypatch):
     # Simulate CLI args like: python run.py /my/path BenchmarkName ModelName --n_problems 10
@@ -14,8 +14,8 @@ def test_run_main(monkeypatch):
     ])
 
     # Patch the Proctor class to avoid real execution
-    with mock.patch("run.Proctor") as MockProctor:
-        run.main()
+    with mock.patch("tether.core.run.Proctor") as MockProctor:
+        main()
         
         # Check that Proctor was called with expected arguments
         MockProctor.assert_called_once()
